Corrected some errors

+++ Data/Image_Label Preprocessing/2_JSON2YOLO_Copy_Move_Corresponding_Image.py
@@ -14,13 +14,13 @@ import cv2
 # ==============================================================
 # 0. 변수 정의
 # ==============================================================
-parser = argparse.ArgumentParser(description='0_2_JSON2YOLO_Copy_Move_Corresponding_Image')
+parser = argparse.ArgumentParser(description='2_JSON2YOLO_Copy_Move_Corresponding_Image')
 
 parser.add_argument('--json-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/annotations', type=str, help='json 파일이 모여있는 폴더 지정')
-parser.add_argument('--json-filenames', default=['instances_train2017.json', 'instances_val2017.json'], type=str, help='yolo로 변환할 json 파일명 지정 (단, json 파일명에 train, val, test 중 해당하는 단어를 포함하고 있어야함)')
-parser.add_argument('--val-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/val2017', type=str, help='val 이미지들이 모여있는 폴더 지정')
-parser.add_argument('--train-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/train2017', type=str, help='train 이미지들이 모여있는 폴더 지정')
-parser.add_argument('--test-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/test2017', type=str, help='test 이미지들이 모여있는 폴더 지정')
+parser.add_argument('--json-filenames', default=['instances_train2017.json', 'instances_val2017.json'], type=str, help='yolo로 변환할 json 파일명 지정 (단, json_mother_abs_path 폴더의 자식으로 파일이 존재해야함)')
+parser.add_argument('--val-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/val2017', type=str, help='val 이미지들이 모여있는 폴더 지정 (단, json 파일명에 train, val, test 중 해당하는 단어를 포함하고 있어야함)')
+parser.add_argument('--train-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/train2017', type=str, help='train 이미지들이 모여있는 폴더 지정 (단, json 파일명에 train, val, test 중 해당하는 단어를 포함하고 있어야함)')
+parser.add_argument('--test-image-mother-abs-path', default='/media/hi/SK Gold P31/Capstone/COCO/test2017', type=str, help='test 이미지들이 모여있는 폴더 지정 (단, json 파일명에 train, val, test 중 해당하는 단어를 포함하고 있어야함)')
 parser.add_argument('--target-cls', default=['37'], type=str, help='yolo 파일로 변환시킬 class 선택, 모든 class를 변환하고 싶은 경우 All 이라고 지정하면 됨')
 
 args = parser.parse_args()
@@ -47,7 +47,7 @@ def main():
         os.makedirs(f"{args.json_mother_abs_path}/images")
 
     # --------------------------------------------------------------
-    # 2) labes, images 각각 train_val_dir 폴더 생성
+    # 2) labes, images 각각 train_val_test_dir 폴더 생성
     # --------------------------------------------------------------
     for train_val_dir in ['train', 'val', 'test']:
         if not os.path.exists(f"{args.json_mother_abs_path}/labels/{train_val_dir}"):
@@ -112,45 +112,50 @@ def main():
 
                     target_image_folder_abs_path = f"{args.json_mother_abs_path}/images/{train_val_dir}"
 
-                    # 4] 이미지 img_x, img_y 추출 + 이미지 정보 추출
-                    image = cv2.imread(image_abs_path)
-                    img_h, img_w, img_c = image.shape
+                    # 4] 이미지가 실제로 존재하는 경우
+                    if os.path.isfile(image_abs_path):
+                        # [1] 이미지 img_x, img_y 추출 + 이미지 정보 추출
+                        image = cv2.imread(image_abs_path)
+                        img_h, img_w, img_c = image.shape
 
-                    # 5] x, y, w, h 추출 & 변환
-                        # [1] x, y의 좌표가 이미지 최소 좌표보다 작은 경우, 0으로 바꿔줌
-                        # [2] x, y 좌표가 실제 w, h 좌표보다 큰 경우, 제거
-                        # [3] w, h좌표가 이미지 실제 w, h보다 큰 경우 이미지 최대 픽셀로 바꿔줌
-                    x, y, w, h = json_data["annotations"][j]["bbox"]
+                        # [2] x, y, w, h 추출 & 변환
+                            # 1]] x, y의 좌표가 이미지 최소 좌표보다 작은 경우, 0으로 바꿔줌
+                            # 2]] x, y 좌표가 실제 w, h 좌표보다 큰 경우, 제거
+                            # 3]] w, h좌표가 이미지 실제 w, h보다 큰 경우 이미지 최대 픽셀로 바꿔줌
+                        x, y, w, h = json_data["annotations"][j]["bbox"]
 
-                    x = max(int(x), 0)
-                    y = max(int(y), 0)
+                        x = max(int(x), 0)
+                        y = max(int(y), 0)
 
-                    if x >= img_w or y >= img_h:
-                        print(f'errors_info : {txt_abs_path} {cls} {x} {y} {img_w} {img_h}')
-                        errors_info.append(f'{txt_abs_path} {cls} {x} {y} {img_w} {img_h}')
+                        if x >= img_w or y >= img_h:
+                            print(f'errors_info : {txt_abs_path} {cls} {x} {y} {img_w} {img_h}')
+                            errors_info.append(f'{txt_abs_path} {cls} {x} {y} {img_w} {img_h}')
 
-                    w = min(int(w), img_w - x)
-                    h = min(int(h), img_h - y)
+                        w = min(int(w), img_w - x)
+                        h = min(int(h), img_h - y)
 
-                    # 6] cx, cy, nw, nh 변환 : 정수 좌표 -> 비율 좌표 변환
-                    cx = (x + w / 2.) / img_w
-                    cy = (y + h / 2.) / img_h
-                    nw = float(w) / img_w
-                    nh = float(h) / img_h
+                        # [3] cx, cy, nw, nh 변환 : 정수 좌표 -> 비율 좌표 변환
+                        cx = (x + w / 2.) / img_w
+                        cy = (y + h / 2.) / img_h
+                        nw = float(w) / img_w
+                        nh = float(h) / img_h
 
-                    line = '%d %.6f %.6f %.6f %.6f\n' % (cls, cx, cy, nw, nh)
+                        line = '%d %.6f %.6f %.6f %.6f\n' % (cls, cx, cy, nw, nh)
 
-                    # 7] yolo 형식 변환 파일 저장
-                    with open(txt_abs_path, 'a') as txt_file:
-                        txt_file.write(line)
+                        # [4] yolo 형식 변환 파일 저장
+                        with open(txt_abs_path, 'a') as txt_file:
+                            txt_file.write(line)
 
-                    # 8] yolo 형식에 맞게 이미지 파일 복사
-                    shutil.copy(image_abs_path, target_image_folder_abs_path)
+                        # [5] yolo 형식에 맞게 이미지 파일 복사
+                        shutil.copy(image_abs_path, target_image_folder_abs_path)
 
-                    # 9] 분류 완료한 이미지 경로 모아두기
-                    images_abs_path.append(image_abs_path)
+                        # [6] 분류 완료한 이미지 경로 모아두기
+                        images_abs_path.append(image_abs_path)
+                    # 5] 이미지가 실제로 존재하지 않는 경우
+                    else:
+                        print(f"No file in local : {image_abs_path}")
 
-                # 9] 출력 : yolo 변환 완료 annotation 개수
+                # 6] 출력 : yolo 변환 완료 annotation 개수
                 if j % ((len(json_data["annotations"]) // 50)+1) == 0:
                     print(f'annotation_count : {j}/{len(json_data["annotations"])}')
 
@@ -165,7 +170,7 @@ def main():
                     set_txt.write('%s\n' % image_abs_path)
 
             # --------------------------------------------------------------
-            # (5) 분류 완료한 이미지 경로 저장 (중복 제거)
+            # (5) 에러 정보 저장
             # --------------------------------------------------------------
             errors_info = set(errors_info)
